fix(*): use `await` instead of `setTimeout` to attach event listeners
