+++ src/components/TemplateLibrary/index.js
@@ -6,9 +6,11 @@ import { TemplateLibrary, Template, Clause } from '@accordproject/cicero-core';
 import { CiceroMarkTransformer } from '@accordproject/markdown-cicero';
 
 import renderNodes from '../../utils/CiceroMarkToOOXML';
+import variableChangeListener from '../../utils/VariableChangeListener';
 
 const LibraryComponent = () => {
   const [templates, setTemplates] = useState(null);
+  const [overallCounter, setOverallCounter] = useState({});
 
   useEffect(() => {
     async function load() {
@@ -24,8 +26,43 @@ const LibraryComponent = () => {
 
   const setup = async dom => {
     await Word.run(async context => {
+      let counter = {};
       dom.nodes.forEach(node => {
-        renderNodes(context, node);
+        renderNodes(context, node, counter);
+      });
+      await context.sync();
+      for (const variable in counter) {
+        for (let count=1; count<=counter[variable]; ++count) {
+          const title = `${variable.toUpperCase()[0]}${variable.substring(1)}${count}`;
+          await attachVariableChangeListener(title);
+        }
+      }
+      setOverallCounter({
+        ...overallCounter,
+        ...counter,
+      });
+    });
+  };
+
+  const attachVariableChangeListener = async title => {
+    await Word.run(async context => {
+      Office.context.document.bindings.addFromNamedItemAsync(title, 'text', { id: title }, res => {
+        if (res.status === Office.AsyncResultStatus.Succeeded) {
+          res.value.addHandlerAsync(Office.EventType.BindingDataChanged, variableChangeListener, res => {
+            if (res.status === Office.AsyncResultStatus.Succeeded) {
+              // ToDo: show the success to user in Production environment
+              console.info(`Listener attached to ${title}`);
+            }
+            else {
+              // ToDo: show the error to user in Production environment
+              console.error(`Listener failed to attach to ${title}`);
+            }
+          });
+        }
+        else {
+          // ToDo: show the error to user in Production environment
+          console.error(title, res);
+        }
       });
       await context.sync();
     });
